warnings

+++ test/PagedVector-test.cpp
@@ -49,14 +49,14 @@ TEST_CASE("PagedVector copy constructor") {
 
 TEST_CASE("PagedVector pass-through") {
   std::vector<double> v(10000);
-  for (int i = 0; i < v.size(); ++i) {
+  for (size_t i = 0; i < v.size(); ++i) {
     v[i] = 2 * i + 1;
   }
   auto v1 = PagedVector<double>(v.data(), v.size());
   auto &w = v1.m_cache.buffer;
   double r = 0;
   bool ad = true;
-  for (int i = 0; i < v.size(); ++i) {
+  for (size_t i = 0; i < v.size(); ++i) {
     ad = ad && (&(w[i]) == &(v[i]));
     r += std::fabs(v[i] - w[i]);
   }
